% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_data_ibague.R
\docType{data}
\name{demographic_data_ibague}
\alias{demographic_data_ibague}
\title{demographic data sample}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 99 rows and 98 columns.
}
\source{
\verb{https://www.dane.gov.co/files/geoportal-provisional/index.html}
}
\usage{
demographic_data_ibague
}
\description{
Demographic data for a given municipality in Colombia (DIVIPOLA
code 73001) from the national census reported in the MGN.
}
\details{
Demographic data sample for testing
}
\keyword{datasets}
