% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{extract_geometry_as_text}
\alias{extract_geometry_as_text}
\title{Extract geometry from shapefile and export it as a character vector}
\usage{
extract_geometry_as_text(geo_data)
}
\arguments{
\item{geo_data}{Shapefile object with geographical data}
}
\value{
Vector with shapefile's geometry as wkt
}
\description{
Extract geometry from shapefile and export it as a character vector
}
\examples{
\dontrun{
extract_geometry_as_text(geo_data)
}
}
\keyword{internal}
