% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{filter_demographic_data}
\alias{filter_demographic_data}
\title{Filter demographic data when imported directly from the MGN}
\usage{
filter_demographic_data(demographic_data)
}
\arguments{
\item{demographic_data}{sf object with demographic data directly imported
from the MGN.}
}
\value{
sf object with filtered data, containing only the necessary columns
for simulation
}
\description{
Filter demographic data when imported directly from the MGN
}
\examples{
\dontrun{
filter_demographic_data(demographic_data)
}
}
