% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{internal_simulation_cpp}
\alias{internal_simulation_cpp}
\title{run simulation}
\usage{
internal_simulation_cpp(
  nDays,
  demographicData,
  temperatureData,
  movementData,
  incubationPeriod,
  infectionDuration,
  initInfectedHumans,
  initInfectedMosquitoes
)
}
\arguments{
\item{nDays}{Number of days to simulate.}

\item{demographicData}{Data frame containing the Demographic data for the
synthetic population.}

\item{temperatureData}{Data frame with the temperature historic.}

\item{movementData}{Data frame with the movement flow between territories.}

\item{incubationPeriod}{Incubation period of the disease.}

\item{infectionDuration}{Infection duration of the disease.}

\item{initInfectedHumans}{Percentage of initially infected humans.}

\item{initInfectedMosquitoes}{Percentage of initially infected mosquitoes.}
}
\value{
Dataframe with susceptible, infected an recovered individuals each
day
}
\description{
run simulation
}
\keyword{internal}
