% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{main}
\alias{main}
\title{run simulation}
\usage{
main(
  nDays,
  demographicData,
  temperatureData,
  movementData,
  incubationPeriod,
  infectionDuration,
  initInfectedHumans,
  initInfectedMosquitoes
)
}
\arguments{
\item{demographicData}{Data frame containing the Demographic data for the
synthetic population.}

\item{temperatureData}{Data frame with the temperature historic.}

\item{movementData}{Data frame with the movement flow between territories.}

\item{incubationPeriod}{Incubation period of the disease.}

\item{infectionDuration}{Infection duration of the disease.}

\item{initInfectedHumans}{Percentage of initially infected humans.}

\item{initInfectedMosquito}{Percentage of initially infected mosquitoes.}
}
\description{
run simulation
}
