% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_data_ibague.R
\docType{data}
\name{movement_data_ibague}
\alias{movement_data_ibague}
\title{movement data sample}
\format{
An object of class \code{data.frame} with 65 rows and 65 columns.
}
\usage{
movement_data_ibague
}
\description{
Movement data for a given municipality in Colombia (DIVIPOLA
code 73001).
}
\details{
Movement data sample for testing
}
\keyword{datasets}
