% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup ABM data and basic parameters}
\usage{
setup(
  demographic_data,
  temperature_data,
  movement_data = NULL,
  demographic_data_MGN = TRUE,
  divipola_code = NULL,
  disease = "dengue",
  incubation_period = 14,
  init_infected_humans = 0.005,
  init_infected_mosquito = 0.03
)
}
\arguments{
\item{demographic_data}{sf object with the demographic data in the specified
format. It's expected to be input by the user from the DANE´s MGN (Marco
Geoestadístico Nacional).}

\item{temperature_data}{Data frame with the temperature historic. It's
expected to be input from the historic data at IDEAM´s historic temperature}

\item{movement_data}{Data frame with the movement flow between territories.
If not provided it's automatically calculated from the demographic data using
distance and population}

\item{demographic_data_MGN}{Boolean stating if the data is directly
downloaded from MGN and needs preparing and cleaning}

\item{divipola_code}{Divipola code of the municipality to be simulated}

\item{disease}{Character string for the disease name. Can be either "dengue",
"zika" or "chikungunya".}

\item{incubation_period}{Incubation period of the disease}

\item{init_infected_humans}{Percentage of initially infected humans}

\item{init_infected_mosquito}{Percentage of initially infected mosquitoes}
}
\value{
S3 ABM_model object with the saved initial parameters
}
\description{
Setup ABM data and basic parameters
}
\examples{
\dontrun{
setup()
}
}
