% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Run simulation for a given number of days.}
\usage{
simulation(ABM_data, time)
}
\arguments{
\item{ABM_data}{S3 class object containing the parameters for the
simulation.}

\item{time}{Integer for the amount of days to simulate.}
}
\value{
Data frame containing SIR simulation results per day.
}
\description{
Run simulation for a given number of days.
}
\examples{
\dontrun{
simulation(ABM_model, 100)
}
}
