% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_data_ibague.R
\docType{data}
\name{temperature_data_ibague}
\alias{temperature_data_ibague}
\title{temperature data sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 2 columns.
}
\source{
\verb{http://www.ideam.gov.co/web/tiempo-y-clima}
}
\usage{
temperature_data_ibague
}
\description{
Temperature data for a given municipality in Colombia (DIVIPOLA
code 73001) between January and June 2023.
}
\details{
Temperature data sample for testing
}
\keyword{datasets}
